allow webm video files to play using videojs-ogvjs
As there is no native support for webm in Webkit, we use a JS implementation
of the codec, called using video.js.

the mechanism is use fails to be detected by webkit as a "user-initiated play" resulting
in webm videos to be paused immediately on play.

To overcome this, we allow all media files to be played without user action.
> previous youtube ZIM files had autoplay enabled.